Fix progress circle and some style
